debt - move supported encodings closer to where they belong
